backend/requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
motor==3.7.1
pymongo==4.16.0
pydantic[email]==2.12.5
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
